\name{HiddenMarkov-defunct}
\alias{HiddenMarkov-defunct}
\alias{residualshmm}

\title{Defunct Functions in Package HiddenMarkov}

\description{
These functions are now defunct and are no longer in the package.
}

\usage{
residualshmm(x, Pi, delta, distn, pm, pn = NULL, discrete = FALSE)
}

\keyword{internal}
